added header file for size_t

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,5 @@
 #include "function_pointers.h"
-#include <stdio.h>
-#include <stddef.h>
+#include<stdlib.h>
 
 /**
  * array_iterator - execute arrays
@@ -14,7 +13,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	int i = 0;
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 	{
 		action(array[i]);
 	}
